# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Real Estate Investment API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/notifications:
    post:
      operationId: broadcastNotification
      tags: [admin]
      summary: Broadcast notification to all users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BroadcastNotificationBody"
      responses:
        "201":
          description: Notification sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
...
  /admin/users/{id}/activate-level:
    patch:
      operationId: activateUserLevel
      tags: [admin]
      summary: Activate or deactivate a position level for a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminActivateLevelBody"
      responses:
        "200":
          description: Level updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserItem"
  /admin/users/{id}:
    patch:
      operationId: updateAdminUser
      tags: [admin]
      summary: Update user account (activate/disable/edit)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserItem"
    delete:
      operationId: deleteAdminUser
      tags: [admin]
      summary: Delete a user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
  /admin/users/{id}/balance-adjust:
    post:
      operationId: adminBalanceAdjust
      tags: [admin]
      summary: Credit or debit a user's balance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminBalanceAdjustBody"
      responses:
        "200":
          description: Balance adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminBalanceAdjustResponse"
...
  /admin/withdrawal-requests/{id}:
    patch:
      operationId: processWithdrawalRequest
      tags: [admin]
      summary: Approve or deny a withdrawal request
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessWithdrawalBody"
      responses:
        "200":
          description: Processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
    delete:
      operationId: deleteWithdrawalRequest
      tags: [admin]
      summary: Delete a withdrawal request (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
...
  /admin/withdrawal-settings:
    get:
      operationId: getWithdrawalSettings
      tags: [admin]
      summary: Get master withdrawal lock settings
      responses:
        "200":
          description: Current withdrawal settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalSettings"
    patch:
      operationId: updateWithdrawalSettings
      tags: [admin]
      summary: Update master withdrawal lock (lock/unlock all users + set timeframe)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWithdrawalSettingsBody"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalSettings"
  /admin/users/{id}/withdrawal-lock:
    patch:
      operationId: toggleUserWithdrawalLock
      tags: [admin]
      summary: Lock or unlock a specific user's withdrawal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleUserWithdrawalLockBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserItem"
...
  /admin/lock-funds-visible:
    get:
      operationId: getAdminLockFundsVisible
      tags: [admin]
      summary: Get Lock Funds card visibility setting (admin)
      responses:
        "200":
          description: Current visibility state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LockFundsVisibleResponse"
    post:
      operationId: setAdminLockFundsVisible
      tags: [admin]
      summary: Enable or disable the Lock Funds card on user dashboards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetLockFundsVisibleBody"
      responses:
        "200":
          description: Visibility updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LockFundsVisibleResponse"
...
  /admin/flash-message:
    get:
      operationId: getAdminFlashMessage
      tags: [admin]
      summary: Get current flash message (admin)
      responses:
        "200":
          description: Current flash message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashMessageResponse"
    post:
      operationId: setFlashMessage
      tags: [admin]
      summary: Set or update the flash message shown to all users on login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetFlashMessageBody"
      responses:
        "200":
          description: Flash message updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashMessageResponse"
    delete:
      operationId: clearFlashMessage
      tags: [admin]
      summary: Clear the flash message
      responses:
        "200":
          description: Flash message cleared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FlashMessageResponse"
components:
  schemas:
...
    SuccessMessage:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
...
    BroadcastNotificationBody:
      type: object
      required: [title, message]
      properties:
        title:
          type: string
        message:
          type: string
    AdminUserItem:
      type: object
      required: [id, firstName, surname, email, phone, whatsappNumber, position, level, role, isActive, balance, referralCode, createdAt, activatedLevels, withdrawalLocked]
      properties:
        id:
          type: integer
        firstName:
          type: string
        surname:
          type: string
        username:
          type: string
        email:
          type: string
        phone:
          type: string
        whatsappNumber:
          type: string
        plainPassword:
          type: string
        position:
          type: string
        level:
          type: string
        role:
          type: string
        isActive:
          type: boolean
        withdrawalLocked:
          type: boolean
        balance:
          type: number
        securityDeposit:
          type: number
        referralCode:
          type: string
        createdAt:
          type: string
        activatedLevels:
          type: array
          items:
            type: string
    WithdrawalSettings:
      type: object
      required: [masterLocked, lockDays]
      properties:
        masterLocked:
          type: boolean
        lockDays:
          type: integer
        lockedAt:
          type: string
          nullable: true
        unlockAt:
          type: string
          nullable: true
        manualLocked:
          type: boolean
        autoScheduleEnabled:
          type: boolean
    UpdateWithdrawalSettingsBody:
      type: object
      required: [masterLocked]
      properties:
        masterLocked:
          type: boolean
        lockDays:
          type: integer
        manualLocked:
          type: boolean
        autoScheduleEnabled:
          type: boolean
    ToggleUserWithdrawalLockBody:
      type: object
      required: [locked]
      properties:
        locked:
          type: boolean
    AdminActivateLevelBody:
      type: object
      required: [levelKey, action]
      properties:
        levelKey:
          type: string
        action:
          type: string
          enum: [activate, deactivate]
    AdminUpdateUserBody:
      type: object
      properties:
        isActive:
          type: boolean
        role:
          type: string
        position:
          type: string
        level:
          type: string
        balance:
          type: number
        securityDeposit:
          type: number
        firstName:
          type: string
        surname:
          type: string
        email:
          type: string
        whatsappNumber:
          type: string
        bankName:
          type: string
        accountNumber:
          type: string
        accountHolderName:
          type: string
...
    ProcessWithdrawalBody:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, denied]
        adminNote:
          type: string
...
    AdminBalanceAdjustBody:
      type: object
      required: [type, amount]
      properties:
        type:
          type: string
          enum: [credit, debit]
        amount:
          type: number
        note:
          type: string
    AdminBalanceAdjustResponse:
      type: object
      required: [success, newBalance]
      properties:
        success:
          type: boolean
        newBalance:
          type: number
...
    LockFundsVisibleResponse:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
    SetLockFundsVisibleBody:
      type: object
      required: [enabled]
      properties:
        enabled:
          type: boolean
    FlashMessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
          nullable: true
    SetFlashMessageBody:
      type: object
      required: [message]
      properties:
        message:
          type: string